requirements.txt
django-qrcode
pandas==2.1.4
qrcode[pil]==7.4.2
openpyxl==3.1.2
pillow==10.2.0
pyinstaller==6.5.0
